Better handling of debug requests
