+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from './errors.types';
 export * from './responses.types';
 export * from './client.types';
 export * from './auth.types';
+export { IGetHttpRequestArgs } from './request-maker.mixin.types';
